// src/types/location.type.ts
export interface LocationType {
  code: string
  fullName: string
}

export interface DistrictType {
  code: string
  codeName: string
  fullName: string
  fullNameEn: string
  name: string
  nameEn: string
  provinceCode: string
}

export interface WardType {
  code: string
  codeName: string
  fullName: string
  fullNameEn: string
  name: string
  nameEn: string
  districtCode: string
}